km-bbb: mkfs need not wait on user input
At the start of the script, the user is shown all block
devices and his chosen device and confirmation is taken
from the user whether to proceed with loading the image
or not. So there is no need to have mkfs wait on user
input for creating a file system on the partition
created.

Signed-off-by: Krishna Konda <kkchaitu@gmail.com>